refactor stream code
