+++ ui/dashboard/src/components/objects/Streamer/index.vue
@@ -41,7 +41,6 @@
       :src="stream.url"
       preload="auto"
       controls
-      @play="handlePlay"
       @ended="object.isPlaying = false"
       style="display: none"
     ></video>
@@ -52,7 +51,8 @@
 import { computed, reactive, ref, watch } from "vue";
 import Object from "../Object.vue";
 import { useStore } from "vuex";
-import { cropImageFromCanvas } from "@/utils/canvas";
+import { useShape } from "./composable";
+
 export default {
   components: { Object },
   props: ["stream"],
@@ -60,58 +60,22 @@ export default {
     const store = useStore();
     const object = reactive({ ...props.stream });
     const video = ref();
-    const canvas = document.createElement("canvas");
-    const ctx = canvas.getContext("2d");
-
-    // watchEffect(() => {
-    //   object.src =
-    //     store.state.stage.board.streams[props.stream.id] ?? object.url;
-    // });
 
     const isHost = computed(() =>
       store.state.stage.hosts.some((stream) => stream.id === props.stream.id)
     );
-    const handlePlay = (e) => {
-      const $this = e.target;
-
-      (function loop() {
-        if (!$this.paused && !$this.ended) {
-          ctx.drawImage($this, 0, 0);
-          const { src } = cropImageFromCanvas(canvas);
-          // store.dispatch("stage/sendStreamData", { id: props.stream.id, src });
-          object.src = src;
-          setTimeout(loop, 1000 / 30); // drawing at 30fps
-        }
-      })();
-    };
-
-    const changeShape = {
-      circle: () => {
-        ctx.save();
-        ctx.beginPath();
-        const r = canvas.height / 2;
-        ctx.arc(r, r, r, 0, Math.PI * 2, true);
-        ctx.clip();
-        ctx.clearRect(0, 0, canvas.width, canvas.height);
-        video.value.play();
-        if (!object.isPlaying) {
-          video.value.pause();
-        }
-      },
-    };
 
     watch(props.stream, () => {
-      console.log(props.stream.src);
       if (props.stream.isPlaying) {
         video.value.play();
       } else {
         video.value.pause();
       }
-      if (object.shape !== props.stream.shape) {
-        changeShape[props.stream.shape]?.apply();
-      }
       window.Object.assign(object, props.stream);
     });
+
+    const { src } = useShape(video.value, object.shape);
+    watch(src, () => (object.src = src.value));
 
     const playStream = ({ closeMenu }) => {
       store
@@ -142,7 +106,6 @@ export default {
       video,
       object,
       isHost,
-      handlePlay,
       playStream,
       pauseStream,
       clipCircle,
